// packages/vreo/resources/typings/VreoUnit.ts
export enum VreoKeyframeEnum {
  /** 运镜 */
  CameraMovement = 'CameraMovement',
  /**
   * 全景标签
   */
  PanoTag = 'PanoTag',
  /**
   * 全景文本标签
   */
  PanoTextLabel = 'PanoTextLabel',
  /**
   * 提词器
   */
  Prompter = 'Prompter',
  /**
   * 更新 VR 全景
   */
  UpdateVRPanorama = 'UpdateVRPanorama',
  /**
   * 视频广告
   */
  ModelVideo = 'ModelVideo',
  /**
   * 全景特效
   */
  PanoEffect = 'PanoEffect',
  /**
   * 信息面板
   */
  InfoPanel = 'InfoPanel',
  /**
   * 视频特效
   */
  VideoEffect = 'VideoEffect',
  /**
   * 背景音乐
   */
  BgMusic = 'BgMusic',
  /**
   * 结束
   */
  Exit = 'Exit',
  /**
   * 自定义剧本帧
   */
  Custom = 'Custom',
}

/**
 * 支持自定义剧本关键帧配置项
 */
export type VreoKeyframeConfigMap = {
  PanoTag?: false
  PanoTextLabel?: false
  Prompter?: false
  InfoPanel?: false
}


/**
 * 剧本关键帧
 */
export interface VreoKeyframe {
  uuid?: string
  type: VreoKeyframeEnum
  start: number
  end: number
  parsed?: boolean
  data: Record<string, any>
}
// ...
export enum InfoPanelTypeEnum {
  Image = 'Image',
  Video = 'Video',
}

export enum InfoPanelStyleEnum {
  Drawer = 'Drawer',
  PopUp = 'PopUp',
}

/**
 * 信息面板
 */
export type InfoPanelData = {
  type: InfoPanelTypeEnum
  style?: InfoPanelStyleEnum
  url: string
  // 标题
  title?: string
  // 副标题
  subTitle?: string
}